Incorporate improvements from crond branch

+++ Dockerfile
@@ -1,33 +1,49 @@
 # Use official PostgreSQL image as base
 ARG POSTGRES_IMAGE=postgres:16
-FROM ${POSTGRES_IMAGE}
+ARG ORIGINAL_ENTRYPOINT=docker-entrypoint.sh
+
+FROM ${POSTGRES_IMAGE} AS base
+#This is here to cachebust this layer for different base images, echos in run
+ARG POSTGRES_IMAGE
 
 # Install dependencies
-RUN set -eux; \
-    apt-get update; \
+RUN <<EOR
+    echo "Base image ${POSTGRES_IMAGE}";
+    set -eux;
+    if [ -d /etc/apt ]; then 
+    apt-get update && 
     apt-get install -y \
         ca-certificates \
-        curl \
-        pgbackrest \
         postgresql-$(pg_config --version | grep -oE '[0-9]+' | head -1)-cron \
         postgresql-$(pg_config --version | grep -oE '[0-9]+' | head -1)-plsh; \
-    rm -rf /var/lib/apt/lists/*
+        pgbackrest;
+    rm -rf /var/lib/apt/lists/*; 
+    elif [ -d /etc/apk ]; then 
+    apk update && 
+    apk add --no-cache \
+        ca-certificates \
+        postgresql-$(pg_config --version | grep -oE '[0-9]+' | head -1)-cron \
+        postgresql-$(pg_config --version | grep -oE '[0-9]+' | head -1)-plsh; \
+        pgbackrest;
+    fi
+EOR
 
-# Create configuration directory
-RUN mkdir -p /etc/pgbackrest/conf.d
+FROM base
+ARG ORIGINAL_ENTRYPOINT
+ENV ENTRYCHAIN="${ORIGINAL_ENTRYPOINT}"
+#Remove this if it exists so entrypoint can honor user volumes and only replace if default
+RUN rm /etc/pgbackrest.conf -f
+RUN mkdir -p /crontabs /tmp/pgbackrest/ /var/log/pgbackrest && chown -R postgres:postgres /tmp/pgbackrest/ /var/log/pgbackrest
 
 # Copy configuration scripts
-COPY entrypoint.sh /usr/local/bin/
-COPY 01-setup-extra.sh /docker-entrypoint-initdb.d/
-COPY 02-post-initialize.sh /docker-entrypoint-initdb.d/
+COPY pgp-entrypoint.sh pgp-healthcheck.sh pgp-update-pgcron.sh /usr/local/bin/
+COPY pgp-init.sh /docker-entrypoint-initdb.d/
 
 # Set permissions
-RUN chmod +x /usr/local/bin/entrypoint.sh \
-    /docker-entrypoint-initdb.d/01-setup-extra.sh \
-    /docker-entrypoint-initdb.d/02-post-initialize.sh
+RUN chmod +x /usr/local/bin/pgp-*.sh \
+    /docker-entrypoint-initdb.d/pgp-init.sh 
 
-RUN chown -R postgres:postgres /var/log/pgbackrest/
+HEALTHCHECK --interval=5m --start-interval=5s --start-period=5m CMD /usr/local/bin/pgp-healthcheck.sh
 
-# Set entrypoint and default command
-ENTRYPOINT ["/usr/local/bin/entrypoint.sh"]
-CMD ["postgres"]
+#This entrypoint will chain the original/next entrypoint from the ENTRYCHAIN var
+ENTRYPOINT ["/usr/local/bin/pgp-entrypoint.sh"]
